+++ composeApp/src/jsMain/kotlin/Main.kt
@@ -22,19 +22,21 @@
 * SOFTWARE.
 *
 */
-package dev.shreyaspatil.capturableExample
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.ComposeViewport
+import dev.shreyaspatil.capturableExample.App
+import kotlinx.browser.document
+import kotlinx.browser.window
+import org.jetbrains.skiko.wasm.onWasmReady
 
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            App()
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    window.onload = {
+        onWasmReady {
+            val body = document.body ?: return@onWasmReady
+            ComposeViewport(body) {
+                App()
+            }
         }
     }
 }
