+++ composeApp/src/jvmMain/kotlin/Main.kt
@@ -22,19 +22,28 @@
 * SOFTWARE.
 *
 */
-package dev.shreyaspatil.capturableExample
+import androidx.compose.desktop.ui.tooling.preview.Preview
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.application
+import androidx.compose.ui.window.rememberWindowState
+import dev.shreyaspatil.capturableExample.App
+import java.awt.Dimension
 
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            App()
-        }
+fun main() = application {
+    Window(
+        title = "Capturable Example",
+        state = rememberWindowState(width = 400.dp, height = 711.dp),
+        onCloseRequest = ::exitApplication
+    ) {
+        window.minimumSize = Dimension(300, 533)
+        App()
     }
+}
+
+@Preview
+@Composable
+fun AppPreview() {
+    App()
 }
