+++ composeApp/src/wasmJsMain/kotlin/Main.kt
@@ -22,19 +22,15 @@
 * SOFTWARE.
 *
 */
-package dev.shreyaspatil.capturableExample
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.ComposeViewport
+import dev.shreyaspatil.capturableExample.App
+import kotlinx.browser.document
 
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            App()
-        }
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    val body = document.body ?: return
+    ComposeViewport(body) {
+        App()
     }
 }
