--- composeApp/src/wasmJsMain/kotlin/main.kt
@@ -1,12 +1,0 @@
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.window.ComposeViewport
-import dev.shreyaspatil.capturableExample.App
-import kotlinx.browser.document
-
-@OptIn(ExperimentalComposeUiApi::class)
-fun main() {
-    val body = document.body ?: return
-    ComposeViewport(body) {
-        App()
-    }
-}
